improvement in ssl context
